Grant or revoke permission to store lockfiles on the user's computer

End users may use this from the command line. Package authors should only call it if they cannot ask for permission interactively using check_ok_to_store(ask = TRUE).

Usage

set_ok_to_store(ok = TRUE)

Arguments

ok

TRUE to grant permission to store lockfiles, FALSE to revoke it and unset options("onetime.dir").

Value

Invisible NULL.

Examples

if (FALSE) { # \dontrun{
set_ok_to_store()
} # }