Run code only once

Onetime allows package authors to run code only once (ever) for a given user. It does so by writing a file, typically to a folder in the user's configuration directory as given by rappdirs::user_config_dir(). The user can set an alternative filepath using options("onetime.dir").

Details

It is package authors' responsibility to check for permission to store lockfiles. This may have been done already by another package if onetime was already installed. You can ask permission interactively on the command line by calling check_ok_to_store() with ask = TRUE.

Core functions include:

onetime_do() runs arbitrary code only once.

onetime_warning() and friends print a warning or message only once.

onetime_message_confirm() prints a message and asks "Show this message again?"

onetime_rlang_warn() and onetime_rlang_inform() print messages using functions from the rlang package.

onetime_only() returns a function that runs only once.

check_ok_to_store() and set_ok_to_store() check for or grant permission to store lockfiles on the user's computer.

Example

library(onetime)
ids  <- paste0("onetime-readme-", sample(1e9, 4))

for (i in 1:5) {
  cat("Loop ", i, " of 5\n")
  onetime_do(cat("You will only see this once.\n"), id = ids[1])
  onetime_warning("This warning will only be shown once.", id = ids[2])
  onetime_message("This message will only be shown once.", id = ids[3])
}

## Loop  1  of 5
## You will only see this once.

## Warning: This warning will only be shown once.

## This message will only be shown once.

## Loop  2  of 5
## Loop  3  of 5
## Loop  4  of 5
## Loop  5  of 5

# Meanwhile, in a separate process:
library(callr)
result <- callr::r(function (ids) {
  onetime::onetime_message("This message with an existing ID will not be shown.", id = ids[1])
  onetime::onetime_message("This message with a new ID will be shown.", id = ids[4])
}, show = TRUE, args = list(ids = ids))

## This message with a new ID will be shown.