Run code only once

This function runs an expression just once. It then creates a lockfile recording a unique ID which will prevent the expression being run again.

Usage

onetime_do(
  expr,
  id = calling_package(),
  path = default_lockfile_dir(),
  expiry = NULL,
  default = NULL,
  without_permission = c("warn", "run", "stop", "pass", "ask")
)

Arguments

expr

The code to evaluate. An R statement or expression() object.

id

Unique ID string. By default, name of the calling package.

path

Directory to store lockfiles.

expiry

difftime() or e.g. lubridate::duration() object. After this length of time, code will be run again.

default

Value to return if expr was not executed.

without_permission

Character string. What to do if the user hasn't given permission to store files? "warn" runs the action with an extra warning; "run" runs the action; "pass" does nothing and returns the default; "stop" throws an error; "ask" asks for permission, after running the action but before recording it on disk.

Value

The value of expr, invisibly; or default if expr was not run because it had been run already.

Details

Calls are identified by id. If you use the same value of id across different calls to onetime_do() and similar functions, only the first call will get made.

By default, id is just the name of the calling package. This is for the common use case of a single call within a package (e.g. at first startup). If you want to use multiple calls, or if the calling code is not within a package, then you must set id explicitly. If you are working in a large project with many contributors, it is strongly recommended to set id explicitly.

The default path, where lockfiles are stored, is within rappdirs::user_config_dir() unless overridden by options("onetime.dir"). If the lockfile cannot be written (e.g. because the user has not given permission to store files on his or her computer), then the call will still be run, so it may be run repeatedly. Conversely, if the call gives an error, the lockfile is still written.

Examples

if (FALSE) {
id <- sample(10000L, 1L)
for (n in 1:3) {
  onetime_do(print("printed once"), id = id)
}

onetime_reset(id = id)
}